Comment final impl

+++ src/main/java/com/project1/project1/services/CommentService.java
@@ -2,11 +2,9 @@ package com.project1.project1.services;
 import com.project1.project1.dto.CommentCreateDto;
 import com.project1.project1.dto.CommentDto;
 import com.project1.project1.dto.ListResponseDto;
-import com.project1.project1.dto.PostFullDto;
 import com.project1.project1.dto.mappers.CommentMapper;
 import com.project1.project1.model.Comment;
 import com.project1.project1.repository.CommentDao;
-import com.project1.project1.repository.PostDao;
 import com.project1.project1.specification.CommentSpecification;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -15,9 +13,6 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.util.List;
-    
-import java.util.Optional;
 import java.util.UUID;
 
 
@@ -37,12 +32,21 @@ public class CommentService {
         return new ListResponseDto<>(CommentPage);
     }
 
+    //Get Comment by post
     public ListResponseDto<CommentDto> getCommentsByPost(UUID id,LocalDate startDate,LocalDate endDate,Pageable pageable) {
         Specification<Comment> spec= CommentSpecification.byPost(id).and(CommentSpecification.filter(startDate,endDate));
         Page<CommentDto> comments=commentDao.findAll(spec,pageable).map(commentMapper::toDto);
         return new ListResponseDto<>(comments);
     }
 
+    //Get Comment by user
+    public ListResponseDto<CommentDto> getCommentsByUser(UUID id,LocalDate startDate,LocalDate endDate,Pageable pageable) {
+        Specification<Comment> spec= CommentSpecification.byUser(id).and(CommentSpecification.filter(startDate,endDate));
+        Page<CommentDto> comments=commentDao.findAll(spec,pageable).map(commentMapper::toDto);
+        return new ListResponseDto<>(comments);
+    }
+
+    //create
     public CommentDto createComment(CommentCreateDto commentDto) {
         if(commentDto.getPost()==null & commentDto.getOwner()==null) {
             throw new IllegalArgumentException("Post Id and User Id must not be null");
@@ -51,7 +55,12 @@ public class CommentService {
         return commentMapper.toDto(commentDao.save(c)) ;
     }
 
-    public void deleteComment(UUID id) {
+    //delete
+    public UUID deleteComment(UUID id) {
+        if(commentDao.findById(id).isEmpty()){
+            throw new IllegalArgumentException("Comment with id: "+id+" does not exist");
+        }
         commentDao.deleteById(id);
+        return id;
     }
 }
